refactor(project): move template library inventory
