+++ crates/djls-project/src/lib.rs
@@ -7,7 +7,6 @@ mod python;
 mod resolve;
 mod settings;
 mod specs;
-mod symbols;
 mod sync;
 mod system;
 mod templates;
@@ -47,7 +46,7 @@ pub use resolve::templatetag_modules;
 pub use settings::django_settings;
 pub use settings::settings_module_file;
 pub use settings::template_dirs;
-pub use settings::template_libraries;
+pub use templates::template_libraries;
 pub use specs::ArgumentCountConstraint;
 pub use specs::AsVar;
 pub use specs::BlockSpec;
@@ -74,12 +73,12 @@ pub use specs::extract_block_specs;
 pub use specs::extract_filter_arities;
 pub use specs::extract_model_graph;
 pub use specs::extract_tag_rules;
-pub use symbols::InstalledSymbolCandidate;
-pub use symbols::InstalledSymbolOrigin;
-pub use symbols::SymbolDefinition;
-pub use symbols::TemplateLibraries;
-pub use symbols::TemplateLibrary;
-pub use symbols::TemplateSymbol;
+pub use templates::InstalledSymbolCandidate;
+pub use templates::InstalledSymbolOrigin;
+pub use templates::SymbolDefinition;
+pub use templates::TemplateLibraries;
+pub use templates::TemplateLibrary;
+pub use templates::TemplateSymbol;
 pub use sync::RefreshData;
 pub use sync::RefreshQuery;
 pub use sync::RefreshQueryResult;
